simplestream-maintainer/cmd_prune: Prevent prunning dangling resources on empty product catalog
If product catalog happens to be empty or missing, entire stream can be wiped out.
If that is really desired, all products and versions can be pruned by simply removing
the entire streams's directory. Therefore, we assume something went wrong, and will
not prune dangling (unreferenced) resources if at least one product is not referenced.

Signed-off-by: Din Music <din.music@canonical.com>