simplestream-maintainer/cmd_prune: Remove todo
First we remove the version from the product catalog to prevent a non-existing version from
being referenced. If removal of the version fails, we log an error. In case that version
we partially removed, it will not be reincluded in the product catalog again, and we will
try to remove it on next prune.

Signed-off-by: Din Music <din.music@canonical.com>