simplestream-maintainer/cmd_build: Ignore versions with mismatched checksums as delta base
When a version with checksum mismatch was detected, it was excluded from the product catalog, however,
delta files were still generated from that version.

This commit ensures that first a valid catalog is constructed (containing only valid product versions)
and only then delta files are calculated based on the catalog instead of the actual directory structure.

Signed-off-by: Din Music <din.music@canonical.com>