simplestream-maintainer: Remove expired product versions
Add prune flag --retain-days which instruct maintainer to remove any product version that is older
then the given number of days, even if the version is referenced in the product catalog.
This way, product versions that are not longer built, for example EOL images, will be automatically
pruned after the given period of time.

Signed-off-by: Din Music <din.music@canonical.com>