simplestream-maintainer/webpage: Parse additional version files from FS
Files are read from the catalog because it already containes their hashes
which are used for displaying image fingerprints that can be used in lxd
to init/launch a specific image version.

However, the webpage builder now also attempts to detect files that are
not present in the catalog. For example, file SHA256SUM.

Signed-off-by: Din Music <din.music@canonical.com>