embed/templates: Dynamically generate image details modal
Image details modals contain a lot of html boilerplate which is repeated
multiple times for each image version. Now, we have only one template that
is populated on the client side when image details are opened.
This reduces the size of non-minified HTML from ~2.3MB to ~500KB (using current
image builds). The majority of size comes from the images data (approximetly
size of the catalog - images.json file).

Signed-off-by: Din Music <din.music@canonical.com>